% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunc.R
\name{betafunc}
\alias{betafunc}
\title{Betafunction for X and y}
\usage{
betafunc(X, y)
}
\arguments{
\item{X}{A matrix with nrow of m.}

\item{y}{A vector with ncol of n.}
}
\value{
The beta's of \code{X} and \code{y}.
}
\description{
Betafunction for X and y
}
\examples{
X = matrix(1:20, ncol = 4, nrow = 5)
y = matrix(c(3,6,4,9,10), ncol = 1)
betafunc(X, y)
}
