% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcars.R
\name{boxcars}
\alias{boxcars}
\title{A function for to make a boxcar for one timing}
\usage{
boxcars(leny, timings, nbox, lenbox, include.intercept = FALSE)
}
\arguments{
\item{leny}{The length of \code{y}.}

\item{timings}{The timing matrix, see \code{\link{timingmatrix}}.}

\item{nbox}{The amount of boxcars in the design matrix.}

\item{lenbox}{The length of each boxcar in the design matrix.}

\item{include.intercept}{Include intercept, default = FALSE.}
}
\value{
The design matrix for one timing with the predefined amount and
  length of boxcars.
}
\description{
A function for to make a boxcar for one timing
}
\examples{
boxcars(100, c(10, 20, 40), nbox = 5, lenbox = 2)
}
\seealso{
\code{\link{timingmatrix}}, for a timing matrix.
}
