% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designmatrix.R
\name{designmatrix}
\alias{designmatrix}
\title{Create a boxcar for all timings}
\usage{
designmatrix(ROI, timings, nbox, lenbox, TR, include.intercept = FALSE)
}
\arguments{
\item{ROI}{A vector of \code{y}.}

\item{timings}{A string of timings.}

\item{nbox}{Amount of boxcars in the design matrix.}

\item{lenbox}{Amount of length of the boxcar in the design matrix.}

\item{TR}{Sampling rate.}

\item{include.intercept}{Include intercept, default is FALSE.}
}
\value{
A combination of \code{\link{timingmatrix}} and \code{\link{boxcars}}.
  This code gives a boxcar for each timing.
}
\description{
Create a boxcar for all timings
}
\examples{
ROI = c(9, -3, -7, -9, -7, -5, -9, 2,  7, 13, 20, 9, 4, 2, 7, 7, -2, -3, -5, 0)
timings = c(2, 6, 8, 10, 12, 14)
nbox = 3
lenbox = 2
TR = 1
designmatrix(ROI, timings, nbox, lenbox, TR)
}
\seealso{
\code{\link{timingmatrix}} for a matrix of timings in which the
  first timing is shuffled, \code{\link{boxcars}}, the boxcar function.
}
