% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbrbetafunc.R
\name{fbrbetafunc}
\alias{fbrbetafunc}
\title{Use the FBR method to find the beta-values}
\usage{
fbrbetafunc(dat, boxtimings, ROI, timings, nbox, lenbox, TR, maximum = TRUE,
  summation = FALSE, all.beta = FALSE)
}
\arguments{
\item{dat}{A nifti file. Can be obtained with readData("filtered_func_data.nii").}

\item{boxtimings}{the designmatrix for each each timing, obtained with \code{\link{designmatrix}}.}

\item{ROI}{The vector of interest \cite{y}.}

\item{timings}{A string of timings.}

\item{nbox}{The amount of boxcars in the design matrix.}

\item{lenbox}{The length of boxcars in the design matrix.}

\item{TR}{Sampling rate.}

\item{maximum}{Maximum is default, other option: summation, all.beta.}

\item{summation}{The sum}

\item{all.beta}{The full beta-matrix}
}
\value{
The beta-values.
}
\description{
Use the FBR method to find the beta-values
}
\examples{
ROI = c(9, -3, -7, -9, -7, -5, -9, 2,  7, 13, 20, 9, 4, 2, 7, 7, -2, -3, -5, 0)
timings = c(2, 6, 8, 10, 12, 14)
nbox = 3
lenbox = 2
TR = 1
boxtimings <- designmatrix(ROI, timings, nbox, lenbox, TR)
dat = 0 #not required in this example, but is readData("filtered_func_data.nii")

fbrbetafunc(dat, boxtimings, ROI, timings, nbox, lenbox, TR, maximum = TRUE )
t(fbrbetafunc(dat, boxtimings, ROI, timings, nbox, lenbox, TR, all.beta = TRUE ))
}
\seealso{
\code{\link{designmatrix}} for the boxcar function for all timings/events.,
  \code{\link[arf3DS4]{readData}} to read nifti files in R
}
