% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.interest.R
\name{main.interest}
\alias{main.interest}
\title{Obtain the beta's with a method of interest}
\usage{
main.interest(boxtimings, ROI, maximum = FALSE, summation = FALSE,
  cancor = FALSE, nbox)
}
\arguments{
\item{boxtimings}{The design matrix for all timings \code{\link{designmatrix}}.}

\item{ROI}{A vector of \code{y}.}

\item{maximum}{Interest in the maximum of all values of beta for each trial.}

\item{summation}{Interest in the summation of all values of beta for each trial.}

\item{cancor}{Interst in the full betas for each box. Note that the matrix
needs to be transposed.}

\item{nbox}{The amount of boxes}
}
\value{
The design matrix for all timings, in which the first timing is shuffled
  in space.
}
\description{
Obtain the beta's with a method of interest
}
\examples{
ROI = c(9, -3, -7, -9, -7, -5, -9, 2,  7, 13, 20, 9, 4, 2, 7, 7, -2, -3, -5, 0)
timings = c(2, 6, 8, 10, 12, 14)
nbox = 3
lenbox = 2
TR = 1
boxtimings <- designmatrix(ROI, timings, nbox, lenbox, TR)

main.interest(boxtimings, ROI, maximum = TRUE, nbox = nbox)
t(main.interest(boxtimings, ROI, cancor = TRUE, nbox = nbox))
}
\seealso{
\code{\link{designmatrix}}, for a boxcar for a timing.
}
