% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timingmatrix.R
\name{timingmatrix}
\alias{timingmatrix}
\title{Shuffle the first timing in a matrix}
\usage{
timingmatrix(timings)
}
\arguments{
\item{timings}{A matrix of the stimulus presentations.}
}
\value{
The design matrix for all timings, in which the first timing is
  shuffled in space.
}
\description{
Shuffle the first timing in a matrix
}
\examples{
timings <- as.matrix(10, 20, 30, 40)
timingmatrix(timings)
}
